+++ src/main/resources/db/migration/V201704091644__add_is_current_event.sql
@@ -1,0 +1,3 @@
+USE `${BREAKOUT}`;
+
+ALTER TABLE `event` ADD `is_current` BOOLEAN DEFAULT FALSE;